support yocto compliance
WRLinux does not use sublayer meta-security-compliance and meta-tpm
in meta-security, but the bbappend of them broke yocto compliance test.

Comment out os-release.bbappend, we have a suspicion we might need this
in a future version and it will be easier to track if it's commented out,
but the bbappend remains.

Remove linux-yocto_4.%.bbappend which we no concerns with this hunk.

Issue: LIN1018-2657

(LOCAL REV: NOT UPSTREAM) -- WR Specific workaround of yocto compliance test

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>